chore: moved URL validation to within Repo Finders
