+++ src/macaron/repo_finder/repo_finder.py
@@ -19,7 +19,7 @@ other configured location).
 For Python, .NET, Rust, and NodeJS type PURLs, Google's Open Source Insights API is used to find the meta data.
 
 In either case, any repository links are extracted from the meta data, then checked for validity via
-``DependencyAnalyzer::find_valid_url`` which accepts URLs that point to a Github repository or similar.
+``utils::find_valid_url`` which accepts URLs that point to a Github repository or similar.
 
 Repository PURLs
 ----------------
@@ -34,7 +34,6 @@ analysis.
 
 import logging
 import os
-from collections.abc import Iterable
 from urllib.parse import ParseResult, urlunparse
 
 from packageurl import PackageURL
@@ -43,7 +42,6 @@ from macaron.config.defaults import defaults
 from macaron.repo_finder.repo_finder_base import BaseRepoFinder
 from macaron.repo_finder.repo_finder_deps_dev import DepsDevRepoFinder
 from macaron.repo_finder.repo_finder_java import JavaRepoFinder
-from macaron.slsa_analyzer.git_url import get_remote_vcs_url
 
 logger: logging.Logger = logging.getLogger(__name__)
 
@@ -77,12 +75,7 @@ def find_repo(purl: PackageURL) -> str:
 
     # Call Repo Finder and return first valid URL
     logger.debug("Analyzing %s with Repo Finder: %s", purl.to_string(), repo_finder.__class__)
-    for urls in repo_finder.find_repo(purl):
-        url = find_valid_url(urls)
-        if url:
-            return url
-
-    return ""
+    return repo_finder.find_repo(purl)
 
 
 def to_domain_from_known_purl_types(purl_type: str) -> str | None:
@@ -155,28 +148,3 @@ def to_repo_path(purl: PackageURL, available_domains: list[str]) -> str | None:
             fragment="",
         )
     )
-
-
-def find_valid_url(urls: Iterable[str]) -> str:
-    """Find a valid URL from the provided URLs.
-
-    Parameters
-    ----------
-    urls : Iterable[str]
-        An Iterable object containing urls.
-
-    Returns
-    -------
-    str
-        A valid URL or empty if it can't find any valid URL.
-    """
-    vcs_set = {get_remote_vcs_url(value) for value in urls if get_remote_vcs_url(value) != ""}
-
-    # To avoid non-deterministic results we sort the URLs.
-    vcs_list = sorted(vcs_set)
-
-    if len(vcs_list) < 1:
-        return ""
-
-    # Report the first valid URL.
-    return vcs_list.pop()
